+++ schematics/src/ng-add/index.d.ts
@@ -1,0 +1,3 @@
+import { Rule } from '@angular-devkit/schematics';
+import { Schema } from './schema';
+export declare function akitaNgAdd(options: Schema): Rule;