+++ schematics/src/ng-add/schema.d.ts
@@ -1,0 +1,11 @@
+export interface Schema {
+    /**
+     * Skip adding dependencies and installing them
+     */
+    skipPackageJson?: boolean;
+    /**
+     * Add the the router store
+     */
+    withRouter?: boolean;
+    project?: any;
+}